+++ saleor/graphql/core/mutations.py
@@ -736,7 +736,17 @@ class ModelMutation(BaseMutation):
         return cls(**{cls._meta.return_field_name: instance, "errors": []})
 
     @classmethod
-    def save(cls, _info: ResolveInfo, instance, _cleaned_input, /):
+    def save(
+        cls,
+        _info: ResolveInfo,
+        instance,
+        _cleaned_input,
+        # Metadata is removed from cleaned input, but sometimes it may be needed in save() method
+        # For example when it saves metadata there
+        metadata_list: list[dict],
+        private_metadata_list: list[dict],
+        /,
+    ):
         instance.save()
 
     @classmethod
@@ -797,7 +807,7 @@ class ModelMutation(BaseMutation):
 
         cls.validate_and_update_metadata(instance, metadata_list, private_metadata_list)
         cls.clean_instance(info, instance)
-        cls.save(info, instance, cleaned_input)
+        cls.save(info, instance, cleaned_input, metadata_list, private_metadata_list)
         cls._save_m2m(info, instance, cleaned_input)
 
         # add to cleaned_input popped metadata to allow running post save events
@@ -866,7 +876,7 @@ class ModelWithRestrictedChannelAccessMutation(ModelMutation):
 
         cls.validate_and_update_metadata(instance, metadata_list, private_metadata_list)
         cls.clean_instance(info, instance)
-        cls.save(info, instance, cleaned_input)
+        cls.save(info, instance, cleaned_input, metadata_list, private_metadata_list)
         cls._save_m2m(info, instance, cleaned_input)
         cls.post_save_action(info, instance, cleaned_input)
         return cls.success_response(instance)